Ready for Production
I discovered that having all the links be founded on an anchor tag
caused the browser to refresh the page every time a button was clicked,
and I also forgot to prevent that default behavior on the link creation
form thus preventing a creation of a new link. I created two different
types of buttons, internal and external, or buttons that take you to
different pages within the site and others that take you to external
sites. I also made the page prettier for desktops and mobile devices.