Almost done...
I'm done working on it for today. I got the editing to work the way I
wanted and got the fun, work, and school pages to work the way I wanted.
I realize I'm copying a lot of code between pages and I want to make it
possible to create new groups of links, so I'll probably make one view
for all link pages, with components that are customizeable. I'll also
have to make big changes to the database and api to make that work (one
collection of all links with a paramter in each document that specifies
what group it should be in. Or learn how to create new collections using
an api...)

Other Ideas for future improvement:
See if I can pull each website's favicon to put in the links.